# docs/swagger.yaml
basePath: /
definitions:
...
  response.Envelope-tenants_TenantListResponse:
    properties:
      data:
        $ref: '#/definitions/tenants.TenantListResponse'
    type: object
  response.Envelope-tenants_TenantResponse:
    properties:
      data:
        $ref: '#/definitions/tenants.TenantResponse'
    type: object
...
  response.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      kind:
        type: string
      message:
        type: string
    type: object
  tenants.CreateTenantRequest:
    properties:
      email:
        example: jordan.lee@example.com
        maxLength: 320
        type: string
      email_opt_in:
        example: true
        type: boolean
      first_name:
        example: Jordan
        maxLength: 120
        minLength: 1
        type: string
      full_name:
        example: Jordan Lee
        maxLength: 512
        type: string
      last_name:
        example: Lee
        maxLength: 120
        type: string
      locale:
        example: en-US
        maxLength: 32
        type: string
      phone:
        example: +1-555-0100
        maxLength: 32
        type: string
      sms_opt_in:
        example: false
        type: boolean
      sms_verified:
        example: false
        type: boolean
      status:
        enum:
        - active
        - inactive
        - archived
        example: active
        type: string
      timezone:
        example: America/Chicago
        maxLength: 64
        type: string
    required:
    - first_name
    type: object
  tenants.CurrentUnitSummary:
    properties:
      unit_id:
        example: 550e8400-e29b-41d4-a716-446655440011
        type: string
      unit_label:
        example: 12B
        type: string
    type: object
  tenants.PatchTenantRequest:
    properties:
      email:
        maxLength: 320
        type: string
      email_opt_in:
        type: boolean
      first_name:
        maxLength: 120
        minLength: 1
        type: string
      full_name:
        maxLength: 512
        type: string
      last_name:
        maxLength: 120
        type: string
      locale:
        maxLength: 32
        type: string
      phone:
        maxLength: 32
        type: string
      sms_opt_in:
        type: boolean
      sms_verified:
        type: boolean
      status:
        enum:
        - active
        - inactive
        - archived
        type: string
      timezone:
        maxLength: 64
        type: string
    type: object
  tenants.TenantListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/tenants.TenantResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
    type: object
  tenants.TenantResponse:
    properties:
      active_lease_count:
        example: 1
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      current_unit:
        $ref: '#/definitions/tenants.CurrentUnitSummary'
      email:
        example: jordan.lee@example.com
        type: string
      email_opt_in:
        type: boolean
      first_name:
        example: Jordan
        type: string
      full_name:
        example: Jordan Lee
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440040
        type: string
      last_name:
        example: Lee
        type: string
      locale:
        example: en-US
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+15550100"
        type: string
      sms_opt_in:
        type: boolean
      sms_verified:
        type: boolean
      status:
        example: active
        type: string
      timezone:
        example: America/Chicago
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Rent Management System HTTP API (bootstrap; module routes are placeholders).
  title: RMS API
  version: "1.0"
paths:
...
  /api/v1/organizations/{id}/tenants:
    get:
      description: '**Nested route:** `/api/v1/organizations/{id}/tenants`. **Admin:**
        any org `id`. **Landlord, manager, staff:** own organization only. **Pagination:**
        `page`, `page_size` (max 100). **Search (AND):** optional `name`, `email`,
        `phone` query params — each matches case-insensitively (`ILIKE`) against name
        fields or email/phone respectively. Optional `status` filter (`active`, `inactive`,
        `archived`). Set `include_summary=true` to add `active_lease_count` and `current_unit`
        (from the latest active lease) per tenant.'
      parameters:
      - description: Organization ID (Gin param name `id`)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: Search by first/last/full name (partial)
        example: Jordan
        in: query
        name: name
        type: string
      - description: Search by email (partial)
        example: lee@
        in: query
        name: email
        type: string
      - description: Search by phone (partial)
        example: "555"
        in: query
        name: phone
        type: string
      - description: Filter by tenant status
        example: active
        in: query
        name: status
        type: string
      - description: true/1/yes to include lease summary fields
        example: "true"
        in: query
        name: include_summary
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Example item: full_name, email_opt_in, active_lease_count
            when include_summary=true'
          schema:
            $ref: '#/definitions/response.Envelope-tenants_TenantListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: '**Admin / landlord / manager** only. Validates `email` format
        when provided; `phone` max length 32. If `full_name` is omitted it is derived
        from first + last name. Default `status` is `active`.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tenants.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-tenants_TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create tenant
      tags:
      - tenants
  /api/v1/organizations/{id}/tenants/{tenantId}:
    delete:
      description: '**Soft-delete** only when the tenant has **no** lease rows (any
        status). If leases exist, returns **409** `tenant_has_leases` — keep the record
        and **PATCH** `status` to `inactive` or `archived` instead. **Admin / landlord
        / manager** only.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        format: uuid
        in: path
        name: tenantId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: tenant_has_leases
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete tenant
      tags:
      - tenants
    get:
      description: Returns a tenant scoped to the organization. Optional `include_summary=true`
        adds `active_lease_count` and `current_unit` from active leases.
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        format: uuid
        in: path
        name: tenantId
        required: true
        type: string
      - description: true/1/yes for lease summary
        example: "true"
        in: query
        name: include_summary
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-tenants_TenantResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get tenant
      tags:
      - tenants
    patch:
      consumes:
      - application/json
      description: '**Preferred lifecycle:** set `status` to `inactive` or `archived`
        instead of deleting when the tenant has lease history. Partial updates supported.
        Recomputes `full_name` when `first_name` or `last_name` changes unless `full_name`
        is sent explicitly.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        format: uuid
        in: path
        name: tenantId
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tenants.PatchTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-tenants_TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update tenant
      tags:
      - tenants
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"